chore(nimbus): reorganize and streamline code
