+++ packages/nimbus/src/components/rich-text-input/utils/is-empty.ts
